refactor out core grid components
